+++ Tests/PlayerTests/Mocks/Common/World/AudioInfoProvider+Mock.swift
@@ -1,9 +1,10 @@
 @testable import Player
 
 extension AudioInfoProvider {
-	static let mock: Self = AudioInfoProvider(
-		isAirPlayOutputRoute: { false },
-		isCarPlayOutputRoute: { false },
-		currentOutputPortName: { "outputPortName" }
-	)
+    static let mock: Self = AudioInfoProvider(
+        isBluetoothOutputRoute: { false },
+        isAirPlayOutputRoute: { false },
+        isCarPlayOutputRoute: { false },
+        currentOutputPortName: { "outputPortName" }
+    )
 }